+++ src/components/molecules/BaseDetail/components/index.tsx
@@ -1,0 +1,4 @@
+import BaseDetailModal from './BaseDetailModal';
+import BaseDetailSwipe from './BaseDetailSwipe';
+
+export {BaseDetailModal, BaseDetailSwipe};