+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.19.0] - 2025-04-29
+
+###
+
+- BaseDetail component
+- ProductDetail component
+
 ## [1.18.0] - 2025-04-28
 
 ### Added